pyproject.toml
[tool.poetry]
name = "agricola-stramlite"
version = "0.1.0"
description = ""
authors = ["Bitboy7 <luisganerx@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit-extras = "^0.5.0"
python-dotenv = "^1.0.1"
mysql-connector-python = "^9.1.0"
openpyxl = "^3.1.5"
extra-streamlit-components = "^0.1.71"
streamlit-lottie = "0.0.5"
st-pages = "0.4.5"
streamlit-modal = "^0.1.2"
streamlit-option-menu = "0.3.12"
streamlit = "1.34.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
